ukl: basic stdio library and lseek, readv and writev
First of all, this patches introduces three new syscalls in UKL: lseek,
readv and writev. All three are needed to implement the following stdio
functions:
- fopen
- fwrite
- fread
- fputs
- fclose
- fflush
- fseek

It is worth noting that, at the moment, none of the stdio API is thread
safe. I've also added wrappers for stdin, stdout and stderr in order to
be compliant with libc behavior.

+++ include/ukl/ukl_internals.h
@@ -4,6 +4,6 @@
 #define UKL_BAD_FD(fn, d)                                                      \
 	printk(KERN_ERR "Trying to call " fn                                   \
 			"() with bad file descriptor: %d",                     \
-	       d)
+	       (int)d)
 
 #endif /* UKL_INTERNALS_H */
